Switch default option if can't save as hybrid
Don't use alert (ugly) or showError (replaces save dialog making it impossible to save).
Mark hybrid option with a warning sign in the dropdown and switch to the next best option when it fails.

Add some safety around showError with eslint.
ESLint doesn't have the best support for this sort of fine-grained restriction. If I had more rules it would replace them - ironically.